Combine __init__.py from types into rule_based_profiler and move everything up one directory. Find and replace rule_based_profiler.types with rule_based_profiler.
