+++ great_expectations/rule_based_profiler/domain_builder/column_domain_builder.py
@@ -3,16 +3,16 @@ from typing import Iterable, List, Optional, Set, Tuple, Union, cast
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.data_context.util import instantiate_class_from_config
-from great_expectations.rule_based_profiler.domain_builder import DomainBuilder
-from great_expectations.rule_based_profiler.helpers.util import (
-    build_domains_from_column_names,
-    get_parameter_value_and_validate_return_type,
-)
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     Domain,
     ParameterContainer,
     SemanticDomainTypes,
     SemanticTypeFilter,
+)
+from great_expectations.rule_based_profiler.domain_builder import DomainBuilder
+from great_expectations.rule_based_profiler.helpers.util import (
+    build_domains_from_column_names,
+    get_parameter_value_and_validate_return_type,
 )
 from great_expectations.validator.metric_configuration import MetricConfiguration
