+++ great_expectations/rule_based_profiler/domain_builder/multi_column_domain_builder.py
@@ -2,13 +2,13 @@ from typing import Dict, List, Optional, Union
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.metric_domain_types import MetricDomainTypes
-from great_expectations.rule_based_profiler.domain_builder import ColumnDomainBuilder
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     INFERRED_SEMANTIC_TYPE_KEY,
     Domain,
     ParameterContainer,
     SemanticDomainTypes,
 )
+from great_expectations.rule_based_profiler.domain_builder import ColumnDomainBuilder
 
 
 class MultiColumnDomainBuilder(ColumnDomainBuilder):
