+++ great_expectations/rule_based_profiler/expectation_configuration_builder/expectation_configuration_builder.py
@@ -5,15 +5,11 @@ from typing import Dict, List, Optional, Set, Union
 from great_expectations.core.batch import Batch, BatchRequestBase
 from great_expectations.core.expectation_configuration import ExpectationConfiguration
 from great_expectations.data_context.util import instantiate_class_from_config
+from great_expectations.rule_based_profiler import Builder, Domain, ParameterContainer
 from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
 from great_expectations.rule_based_profiler.parameter_builder import (
     ParameterBuilder,
     init_rule_parameter_builders,
-)
-from great_expectations.rule_based_profiler.types import (
-    Builder,
-    Domain,
-    ParameterContainer,
 )
 
 logger = logging.getLogger(__name__)
