+++ great_expectations/rule_based_profiler/helpers/configuration_reconciliation.py
@@ -3,10 +3,10 @@ from dataclasses import asdict, dataclass
 from enum import Enum
 
 from great_expectations.core.util import convert_to_json_serializable, nested_update
+from great_expectations.rule_based_profiler import ParameterContainer
 from great_expectations.rule_based_profiler.helpers.util import (
     convert_variables_to_dict,
 )
-from great_expectations.rule_based_profiler.types import ParameterContainer
 from great_expectations.types import SerializableDictDot
 
 logger = logging.getLogger(__name__)
