+++ great_expectations/rule_based_profiler/helpers/simple_semantic_type_filter.py
@@ -2,7 +2,7 @@ from typing import Any, Dict, List, Optional, Union
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.profiler_types_mapping import ProfilerTypeMapping
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     InferredSemanticDomainType,
     SemanticDomainTypes,
     SemanticTypeFilter,
