+++ great_expectations/rule_based_profiler/parameter_builder/metric_single_batch_parameter_builder.py
@@ -1,18 +1,18 @@
 from typing import Dict, List, Optional, Union
 
+from great_expectations.rule_based_profiler import (
+    FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
+    FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
+    Domain,
+    ParameterContainer,
+    ParameterNode,
+)
 from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
 from great_expectations.rule_based_profiler.helpers.util import (
     get_parameter_value_and_validate_return_type,
 )
 from great_expectations.rule_based_profiler.parameter_builder import (
     MetricMultiBatchParameterBuilder,
-)
-from great_expectations.rule_based_profiler.types import (
-    FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
-    FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
-    Domain,
-    ParameterContainer,
-    ParameterNode,
 )
 from great_expectations.types.attributes import Attributes
