+++ great_expectations/rule_based_profiler/parameter_builder/partition_parameter_builder.py
@@ -2,14 +2,7 @@ from typing import Dict, List, Optional, Set, Union
 
 import numpy as np
 
-from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
-from great_expectations.rule_based_profiler.helpers.util import (
-    get_parameter_value_and_validate_return_type,
-)
-from great_expectations.rule_based_profiler.parameter_builder import (
-    MetricSingleBatchParameterBuilder,
-)
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     DOMAIN_KWARGS_PARAMETER_FULLY_QUALIFIED_NAME,
     FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
     FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
@@ -18,6 +11,13 @@ from great_expectations.rule_based_profiler.types import (
     MetricValue,
     ParameterContainer,
     ParameterNode,
+)
+from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
+from great_expectations.rule_based_profiler.helpers.util import (
+    get_parameter_value_and_validate_return_type,
+)
+from great_expectations.rule_based_profiler.parameter_builder import (
+    MetricSingleBatchParameterBuilder,
 )
 from great_expectations.types.attributes import Attributes
