+++ great_expectations/rule_based_profiler/parameter_builder/regex_pattern_string_parameter_builder.py
@@ -2,12 +2,7 @@ import logging
 from typing import Dict, Iterable, List, Optional, Set, Union
 
 import great_expectations.exceptions as ge_exceptions
-from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
-from great_expectations.rule_based_profiler.helpers.util import (
-    get_parameter_value_and_validate_return_type,
-)
-from great_expectations.rule_based_profiler.parameter_builder import ParameterBuilder
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
     FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
     AttributedResolvedMetrics,
@@ -16,6 +11,11 @@ from great_expectations.rule_based_profiler.types import (
     MetricValues,
     ParameterContainer,
 )
+from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
+from great_expectations.rule_based_profiler.helpers.util import (
+    get_parameter_value_and_validate_return_type,
+)
+from great_expectations.rule_based_profiler.parameter_builder import ParameterBuilder
 from great_expectations.types.attributes import Attributes
 
 logger = logging.getLogger(__name__)
