+++ great_expectations/rule_based_profiler/parameter_container.py
@@ -16,7 +16,7 @@ from pyparsing import (
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.util import convert_to_json_serializable
-from great_expectations.rule_based_profiler.types import Domain
+from great_expectations.rule_based_profiler import Domain
 from great_expectations.types import SerializableDictDot, SerializableDotDict
 
 FULLY_QUALIFIED_PARAMETER_NAME_DELIMITER_CHARACTER: str = "$"
