+++ great_expectations/rule_based_profiler/rule/rule_output.py
@@ -1,15 +1,15 @@
 from typing import Dict, List, Optional
 
 from great_expectations.core import ExpectationConfiguration
-from great_expectations.rule_based_profiler.expectation_configuration_builder import (
-    ExpectationConfigurationBuilder,
-)
-from great_expectations.rule_based_profiler.types import (
+from great_expectations.rule_based_profiler import (
     Domain,
     ParameterNode,
     RuleState,
     get_fully_qualified_parameter_names,
     get_parameter_values_for_fully_qualified_parameter_names,
+)
+from great_expectations.rule_based_profiler.expectation_configuration_builder import (
+    ExpectationConfigurationBuilder,
 )
