+++ great_expectations/rule_based_profiler/semantic_type_filter.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Dict, List, Optional, Union
 
-from great_expectations.rule_based_profiler.types import SemanticDomainTypes
+from great_expectations.rule_based_profiler import SemanticDomainTypes
 
 
 class SemanticTypeFilter(ABC):
